Big rename
playground -> tuf-on-ci

* variable and class names
* directories
* packages/modules
* the signer configuration file
* custom metadata field names

This is very incompatible with any existing repositories.
It may be possible to create a playground2tuf-on-ci import script but
this is not included here.

Also remove the downloader client code for now -- We should have an
example but
A) it should be clearly an example, not part of the code release
B) the old code had a lot of unrelated things