Change how key lookup works for "known good" repo
During import we can't connect signers to keys for the "known good"
delegator (because the known good metadata does not have the custom
fields we need). Don't return keys in this case.

This looks a bit iffy but is not security critical:
_get_keys() is used to document the signing status, not to validate
the metadata: that is still done using normal TUF processes.