signer: allow _get_keys() when role does not exist yet
Return an empty list if we are just about to create the role
(this is needed to figure out if invites are required).

Also as a special case for known good keys (i.e. previous root keys):
Skip keys that have now known owner. This allows the import process to
work even though "previous" key owner cannot be figured out.