signer: Fix a special case in import
If the metadata uses "x-playground-" variables and does not contain
x-playground-signing-period for online roles, the import fails
because of order of operations (expiry period is already in new format
when signing period parsing tries to use the old expiry format).

Figure out signing-period before expiry-period so this special case
works.