repo: Start using get_verification_result() from tuf
This allows us to _not_ do some fiddling with thresholds etc.

On the other hand it makes it a little trickier to actually report
thresholds etc to user... I've not even tried that now as the output
was already too fiddly in the root update case: Currently thresholds
are not explicitly printed, only whether more signatures are needed.