actions: Ensure no dispatch when using custom token
The manual workflow dispatches in online-sign and create-signing-events
are only there because GitHub prevents pushes from triggering other
workflows if they are made with GITHUB_TOKEN.

If we use a custom token, skip dispatching n create-signing-events and
assume the push triggers the signing-event workflow automatically.
Publishing is already not triggered by a push (only by dispatch) so
online-sign is fine as is.

Note that online-sign may trigger another online-sign (if it makes a
commit): this seems unavoidable but should not lead to a online-sign
avalanche as long as signing period is not constantly on.