repo: Separate "status" and "update-targets"
This makes the code more logical, but more importantly
it gives us two separate return values
* tuf-on-ci-update-targets return value tells if metadata changes were made
* tuf-on-ci-status return value tells if signing event is complete

This separation is required to handle both the custom token case and the
default token case (former auto-dispatches on every git change, latter
does not so we need logic to manually dispatch in when we do metadata
changes).

Note that signing-event action has not been updated yet in this commit.