actions: Fix test-repository source comparison
Previously setting the initial root would only work if compare_source
was False (because compare_source would lead to a checkout that would
clean the initial root.json that had been put into working dir).

* Change tuf-on-ci-test-client so that --compare-source has an argument
  (the metadata directory to compare against)
* Checkout into a subdirectory: this allows root.json to exist in
  workdir
* Pass the metadata directory to tuf-on-ci-test-client --compare-source

This changes tuf-on-ci-test-client CLI but does not change the action
API so existing users are fine.

Fixes #345