Added support for delegations using cloud KMS
This is mainly done by modifying tuf-on-ci-delegate to accept
an online key as a signer for delegations. This require some extra care
when dealing with the signer's identity. Keyowner or online-uri is now
used.

Tuf-on-ci-online-sign now supports an extra parameter --role. This is
used to perform online signing with delegations.

Signed-off-by: Fredrik Skogman <kommendorkapten@github.com>